Mountains/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
# certifi==2024.2.2
# cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.2
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
executing==2.0.1
Faker==22.0.0
hiredis==2.3.2
idna==3.6
ipython==8.20.0
jedi==0.19.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.4
Unidecode==1.3.8
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1

gevent==24.11.1
gunicorn==23.0.0
python-dotenv==1.1.0
